Hopefully this longer message will work.
I only finished the address and customer with validations. It reads
from the customers file and zip database file and performs validation
on the inputs (although I just remembered I forgot to add a different
message for the zip code, like one similar to the e-mail, so if it's an
incorrect one in the HashMap it'll probably just return null).  The
other classes are varying degrees of finished.  I'm mostly just hoping
for some partial credit since I didn't get much done while I've been
sick.